reorganize loader into separate file
